+++ src/lib.rs
@@ -33,7 +33,9 @@ pub use leafwing_input_manager_macros::Actionlike;
 /// Everything you need to get started
 pub mod prelude {
     pub use crate::action_state::{ActionState, ActionStateDriver};
-    pub use crate::axislike::{DualAxis, MouseWheelAxisType, SingleAxis, VirtualDPad};
+    pub use crate::axislike::{
+        DeadZoneShape, DualAxis, MouseWheelAxisType, SingleAxis, VirtualDPad,
+    };
     pub use crate::buttonlike::MouseWheelDirection;
     pub use crate::clashing_inputs::ClashStrategy;
     pub use crate::input_map::InputMap;