+++ tests/mouse_motion.rs
@@ -103,6 +103,7 @@ fn mouse_motion_dual_axis_mocking() {
             negative_low: 0.0,
             inverted: false,
         },
+        deadzone: DualAxis::DEFAULT_DEADZONE_SHAPE,
     };
     app.send_input(input);
     let mut events = app.world.resource_mut::<Events<MouseMotion>>();